% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_suggested}
\alias{check_suggested}
\title{Check that the version of an suggested package satisfies the requirements}
\usage{
check_suggested(package, version = NULL, compare = NA,
  path = inst("pkgload"))
}
\arguments{
\item{package}{The name of the suggested package}

\item{version}{The version of the package}

\item{compare}{The comparison operator to use to check the version}
}
\description{
Check that the version of an suggested package satisfies the requirements
}
\keyword{internal}
