% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-example.r
\name{dev_example}
\alias{dev_example}
\title{Run a examples for an in-development function.}
\usage{
dev_example(topic, quiet = FALSE)
}
\arguments{
\item{topic}{Name or topic (or name of Rd) file to run examples for}

\item{quiet}{If \code{TRUE}, runs example quietly.}
}
\description{
Run a examples for an in-development function.
}
\examples{
\dontrun{
# Runs installed example:
library("ggplot2")
example("ggplot")

# Runs develoment example:
dev_example("ggplot")
}
}
\seealso{
Other res: \code{\link{compiler_flags}}
}

