% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-index.r
\name{find_pkg_topic}
\alias{find_pkg_topic}
\title{Find help topic in current package}
\usage{
find_pkg_topic(topic, pkg = ".")
}
\arguments{
\item{topic}{name}
}
\description{
Find help topic in current package
}
\keyword{internal}

