% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports-env.r
\name{imports_env}
\alias{imports_env}
\title{Return imports environment for a package}
\usage{
imports_env(package)
}
\arguments{
\item{path}{TODO: fix doc}
}
\description{
Contains objects imported from other packages. Is the parent of the
package namespace environment, and is a child of \url{namespace:base},
which is a child of R_GlobalEnv.
}
\seealso{
\code{\link[=ns_env]{ns_env()}} for the namespace environment that
all the objects (exported and not exported).

\code{\link[=pkg_env]{pkg_env()}} for the attached environment that contains
the exported objects.
}
\keyword{internal}
