% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{is_dev_package}
\alias{is_dev_package}
\title{Is the package currently under development?}
\usage{
is_dev_package(name)
}
\arguments{
\item{name}{the name of a package.}
}
\description{
Returns \code{TRUE} or \code{FALSE} depending on if the package has been loaded by
\strong{pkgload}.
}
