% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_data}
\alias{load_data}
\title{Load data.}
\usage{
load_data(path = ".")
}
\arguments{
\item{path}{Path to a package, or within a package.}
}
\description{
Loads all \code{.RData} files in the data subdirectory.
}
\keyword{programming}
