% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-dll.R
\name{load_dll}
\alias{load_dll}
\title{Load a compiled DLL}
\usage{
load_dll(path = ".")
}
\arguments{
\item{path}{Path to a package, or within a package.}
}
\description{
Load a compiled DLL
}
\keyword{programming}
