% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace-env.r
\name{ns_env}
\alias{ns_env}
\title{Return the namespace environment for a package.}
\usage{
ns_env(package)
}
\arguments{
\item{package}{package name.}
}
\description{
Contains all (exported and non-exported) objects, and is a descendant of
\code{R_GlobalEnv}. The hierarchy is \verb{<namespace:pkg>},
\verb{<imports:pkg>}, \verb{<namespace:base>}, and then
\code{R_GlobalEnv}.
}
\details{
If the package is not loaded, this function returns \code{NULL}.
}
\seealso{
\code{\link[=pkg_env]{pkg_env()}} for the attached environment that
contains the exported objects.

\code{\link[=imports_env]{imports_env()}} for the environment that contains
imported objects for the package.
}
\keyword{internal}
