% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace-env.r
\name{parse_ns_file}
\alias{parse_ns_file}
\title{Parses the NAMESPACE file for a package}
\usage{
parse_ns_file(path = ".")
}
\arguments{
\item{path}{Path to a package, or within a package.}
}
\description{
Parses the NAMESPACE file for a package
}
\examples{
if (has_tests()) {
parse_ns_file(pkgtest("testLoadHooks"))
}
}
\keyword{internal}
