% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgload-package.R
\docType{package}
\name{pkgload-package}
\alias{pkgload}
\alias{pkgload-package}
\title{pkgload: Simulate Package Installation and Attach}
\description{
Simulates the process of installing a package and then attaching it. This is a key part of the 'devtools' package as it allows you to rapidly iterate while developing a package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/pkgload}
  \item \url{https://pkgload.r-lib.org}
  \item Report bugs at \url{https://github.com/r-lib/pkgload/issues}
}

}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@posit.co}

Authors:
\itemize{
  \item Hadley Wickham
  \item Winston Chang
  \item Jim Hester
}

Other contributors:
\itemize{
  \item Mauricio Vargas Sepulveda (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [contributor]
  \item Posit Software, PBC [copyright holder, funder]
  \item R Core team (Some namespace and vignette code extracted from base R) [contributor]
}

}
\keyword{internal}
