% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_tests.R
\name{pkgtest}
\alias{pkgtest}
\title{Return the path to one of the packages in the devtools test dir}
\usage{
pkgtest(package)
}
\arguments{
\item{package}{Name of the test package.}
}
\description{
Devtools comes with some simple packages for testing. This function
returns the path to them.
}
\examples{
if (has_tests()) {
pkgtest("testData")
}
}
\keyword{internal}
