% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-loadhooks.R
\name{run_pkg_hook}
\alias{run_pkg_hook}
\alias{run_user_hook}
\title{Run user and package hooks.}
\usage{
run_pkg_hook(package, hook)

run_user_hook(package, hook)
}
\arguments{
\item{package}{package name.}

\item{hook}{hook name: one of "load", "unload", "attach", or "detach"}
}
\description{
Run user and package hooks.
}
\keyword{internal}
